+++ ferry/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.4]
+
+- catch Stream error events in `ErrorTypedLink`
+- include `ErrorTypedLink` in the default client `TypedLink` chain
+- update dependencies
+
 ## [0.9.3]
 
 - fix bug causing `FetchPolicy.CacheFirst` & `FetchPolicy.CacheAndNetwork` to only receive the first response from Link